tried an extra validation
Want a field that is a positive integer (excluding 0).
Added a new field with a min value.
This validation is at the model level.

Validation is respected.
But now two kinds of validation error messages appear in the form.
The nicely-formatted ones I saw before turn out to be imposed by html5.
The new one, imposed at the model level, is an ugly bullet list above the
field.