template adjustment to how each field is displayed
Before, the template said, "render the whole form."

Now, the template says, "loop through the fields, and do each field like
this."  Getting all the fields is still automatic.

This change allowed putting the error message after the field, and
styling the error message a bit.  It also allows changing the help text
position, if desired.