first working case of crispy formhelper
Followed the end of this tutorial, but for my traits:
https://simpleisbetterthancomplex.com/tutorial/2018/08/13/how-to-use-bootstrap-4-forms-with-django.html

An important component missing from that tutorial was urls.py.
Needed to add a new url for update, of course.
But that url also requires a code for which trait record to update.
Found a solution here:
https://stackoverflow.com/questions/23154525/django-generic-detail-view-must-be-called-with-either-an-object-pk-or-a-slug
There might be a cleaner way to do it.  Could refer to the django-girls blog app, which uses <int:pk>.

Now, can update existing trait records at, e.g.,
http://127.0.0.1:8000/traits/update/4/