adjust slider steps; slider behavior notes
Note that after using the slider to filter, even when the slider is
moved back to the min and max values, records are not shown if they lack
a value.

To unset all filters, click the "N total" link near the search button.
Not clear how to unset just the numeric filter, though.